Stub scanviz command
