+++ cli/src/scanviz/__init__.py
@@ -1,0 +1,1 @@
+from .scanviz import scanviz
